[refactor] Move all configure-related methods to a function (#6951)
* move all configure-related methods to a function

* move all configure-related methods to a function

* typo
