+++ conans/client/graph/graph_builder.py
@@ -1,12 +1,12 @@
 import time
 
+from conans.client.conanfile.configure import run_configure_method
 from conans.client.graph.graph import DepsGraph, Node, RECIPE_EDITABLE, CONTEXT_HOST, CONTEXT_BUILD
 from conans.errors import (ConanException, ConanExceptionInUserConanfileMethod,
                            conanfile_exception_formatter)
 from conans.model.conan_file import get_env_context_manager
 from conans.model.ref import ConanFileReference
 from conans.model.requires import Requirements, Requirement
-from conans.util.conan_v2_mode import conan_v2_behavior
 from conans.util.log import logger
 
 
@@ -321,27 +321,9 @@ class DepsGraphBuilder(object):
         """
         conanfile, ref = node.conanfile, node.ref
         try:
-            # Avoid extra time manipulating the sys.path for python
+            run_configure_method(conanfile, down_options, down_ref, ref)
+
             with get_env_context_manager(conanfile, without_python=True):
-                if hasattr(conanfile, "config"):
-                    conan_v2_behavior("config() has been deprecated. "
-                                      "Use config_options() and configure()",
-                                      v1_behavior=conanfile.output.warn)
-                    with conanfile_exception_formatter(str(conanfile), "config"):
-                        conanfile.config()
-                with conanfile_exception_formatter(str(conanfile), "config_options"):
-                    conanfile.config_options()
-                conanfile.options.propagate_upstream(down_options, down_ref, ref)
-                if hasattr(conanfile, "config"):
-                    with conanfile_exception_formatter(str(conanfile), "config"):
-                        conanfile.config()
-
-                with conanfile_exception_formatter(str(conanfile), "configure"):
-                    conanfile.configure()
-
-                conanfile.settings.validate()  # All has to be ok!
-                conanfile.options.validate()
-
                 # Update requirements (overwrites), computing new upstream
                 if hasattr(conanfile, "requirements"):
                     # If re-evaluating the recipe, in a diamond graph, with different options,
